# web/openapi.json
{
  "components": {
    "schemas": {
...
      "InstallRequest": {
        "properties": {
          "external_id": {
            "description": "The source's own id, as the repository index lists it.",
            "type": "string"
          },
          "repo_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["repo_id", "external_id"],
        "type": "object"
      },
      "InstalledSourceDto": {
        "properties": {
          "declared_rate_limit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RateLimitDto",
                "description": "Requests per period the source asked for, if it has declared one."
              }
            ]
          },
          "external_id": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "repo_id": {
            "format": "uuid",
            "type": "string"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "version": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "repo_id",
          "external_id",
          "name",
          "version",
          "languages",
          "required_capabilities"
        ],
        "type": "object"
      },
      "JobDto": {
        "properties": {
          "attempts": {
            "format": "int32",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "last_error": {
            "type": ["string", "null"]
          },
          "max_attempts": {
            "format": "int32",
            "type": "integer"
          },
          "priority": {
            "format": "int32",
            "type": "integer"
          },
          "run_at": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "priority",
          "run_at",
          "attempts",
          "max_attempts",
          "created_at"
        ],
        "type": "object"
      },
...
      "RateLimitDto": {
        "properties": {
          "period_seconds": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "permits": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["permits", "period_seconds"],
        "type": "object"
      },
      "SettingRequest": {
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "description": "Base64 of the postcard-encoded value. Absent clears the setting.",
            "type": ["string", "null"]
          }
        },
        "required": ["key"],
        "type": "object"
      },
      "SourceEntryDto": {
        "properties": {
          "base_url": {
            "type": ["string", "null"]
          },
          "content_rating": {
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "icon_url": {
            "type": ["string", "null"]
          },
          "installed_version": {
            "description": "Absent when the entry is not installed.",
            "format": "int32",
            "minimum": 0,
            "type": ["integer", "null"]
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "repo_id": {
            "format": "uuid",
            "type": "string"
          },
          "version": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["repo_id", "external_id", "name", "version", "languages", "content_rating"],
        "type": "object"
      },
      "SourceRepoDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "last_refreshed_at": {
            "format": "date-time",
            "type": ["string", "null"]
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["id", "name", "url", "created_at"],
        "type": "object"
      },
      "SourceRepoRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "description": "The index URL, such as `https://example.test/index.min.json`.",
            "type": "string"
          }
        },
        "required": ["name", "url"],
        "type": "object"
      },
      "SourceSettingDto": {
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "description": "Base64 of the postcard-encoded value, or absent when unset.",
            "type": ["string", "null"]
          }
        },
        "required": ["key"],
        "type": "object"
      },
      "SourceSettingsDto": {
        "properties": {
          "declared": {
            "description": "The package's `settings.json`, verbatim. A client renders it; this\nserver does not interpret it."
          },
          "values": {
            "description": "What the source has actually stored, which is not necessarily what the\ndeclaration lists: a source writes keys of its own choosing.",
            "items": {
              "$ref": "#/components/schemas/SourceSettingDto"
            },
            "type": "array"
          }
        },
        "required": ["declared", "values"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "An opaque, signed, HttpOnly session cookie issued by `GET /api/v1/auth/callback`. There is no bearer token: tokens from the identity provider never reach the browser.",
        "in": "cookie",
        "name": "id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted manga library server.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "nyuka",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "description": "Job id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such job"
          }
        },
        "summary": "`GET /api/v1/jobs/{id}`",
        "tags": ["jobs"]
      }
    },
...
    "/source-repos": {
      "get": {
        "operationId": "listSourceRepos",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRepoDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "`GET /api/v1/source-repos`",
        "tags": ["sources"]
      },
      "post": {
        "operationId": "addSourceRepo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRepoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRepoDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Invalid URL"
          }
        },
        "summary": "`POST /api/v1/source-repos`",
        "tags": ["sources"]
      }
    },
    "/source-repos/{id}": {
      "delete": {
        "operationId": "deleteSourceRepo",
        "parameters": [
          {
            "description": "Repository id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed, or already absent"
          }
        },
        "summary": "`DELETE /api/v1/source-repos/{id}`",
        "tags": ["sources"]
      }
    },
    "/source-repos/{id}/available": {
      "get": {
        "operationId": "listAvailableSources",
        "parameters": [
          {
            "description": "Repository id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceEntryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such repository"
          }
        },
        "summary": "`GET /api/v1/source-repos/{id}/available`",
        "tags": ["sources"]
      }
    },
    "/source-repos/{id}/refresh": {
      "post": {
        "operationId": "refreshSourceRepo",
        "parameters": [
          {
            "description": "Repository id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "A refresh was queued"
          },
          "404": {
            "description": "No such repository"
          }
        },
        "summary": "`POST /api/v1/source-repos/{id}/refresh`",
        "tags": ["sources"]
      }
    },
    "/sources": {
      "get": {
        "operationId": "listSources",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstalledSourceDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "`GET /api/v1/sources`",
        "tags": ["sources"]
      },
      "post": {
        "description": "Synchronous rather than a job. The client is waiting to use the source, the\nwork is one download and one compile, and the capability refusal has to\nreach the person who pressed install — a failed job they have to go and\nread reports a host gap as though it were a site problem, which ADR-0004\nnames as the worst outcome.",
        "operationId": "installSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstalledSourceDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such repository or entry"
          },
          "501": {
            "description": "The source needs a capability this build lacks"
          },
          "502": {
            "description": "The package could not be fetched"
          }
        },
        "summary": "`POST /api/v1/sources` — install from a repository entry.",
        "tags": ["sources"]
      }
    },
    "/sources/{id}": {
      "delete": {
        "description": "> [!WARNING]\n> This cascades. Every series from this source and everything hanging off\n> them is removed from the database. Packaged files in the library are left\n> alone — they are the user's — but nothing will know how to find them\n> again.",
        "operationId": "uninstallSource",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed, or already absent"
          }
        },
        "summary": "`DELETE /api/v1/sources/{id}`",
        "tags": ["sources"]
      }
    },
...
    "/sources/{id}/filters": {
      "get": {
        "operationId": "getSourceFilters",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "The source's filter definitions"
          },
          "404": {
            "description": "No such source"
          }
        },
        "summary": "`GET /api/v1/sources/{id}/filters`",
        "tags": ["sources"]
      }
    },
    "/sources/{id}/settings": {
      "get": {
        "description": "# Values are opaque, and that is not laziness\n\nA source's settings are read and written by its WASM module through the\n`defaults` host import, which postcard-encodes them. This server does not\nknow what a given key means to a given source — the shape is declared in\nthe package's `settings.json` and interpreted by the module.\n\nSo values go over the wire base64-encoded rather than decoded into JSON.\nDecoding would mean this server guessing at a schema it does not have, and\nguessing wrong writes a value the source then misreads.",
        "operationId": "getSourceSettings",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSettingsDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such source"
          }
        },
        "summary": "`GET /api/v1/sources/{id}/settings`",
        "tags": ["sources"]
      },
      "put": {
        "operationId": "putSourceSetting",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Stored"
          },
          "400": {
            "description": "Not valid base64, or too large"
          },
          "404": {
            "description": "No such source"
          }
        },
        "summary": "`PUT /api/v1/sources/{id}/settings`",
        "tags": ["sources"]
      }
    }
  },
...
}